src/schoology-extractor/pyproject.toml
[tool.poetry]
name = "edfi-schoology-extractor"
version = "1.0.0-alpha.1"
homepage = "https://techdocs.ed-fi.org/display/EDFITOOLS/LMS+Toolkit"
repository = "https://github.com/Ed-Fi-Exchange-OSS/LMS-Toolkit"
description = "Extract tool for retrieving student data from Canvas"
readme = "README.md"
authors = ["Ed-Fi Alliance, LLC, and contributors"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
pandas = "^1.1.2"
requests_oauthlib = "^1.3.0"
oauthlib = "^3.1.0"
configargparse = "^1.2.3"
python-dotenv = "^0.14.0"
opnieuw = "^1.1.0"
SQLAlchemy = "^1.3.20"
errorhandler = "^2.0.1"
edfi-lms-extractor-lib = {version = "1.0.0a3", allow-prereleases = true}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
black = "^20.8b1"
rope = "^0.17.0"
flake8 = "^3.8.3"
pytest-mock = "^3.3.1"
pyfakefs = "^4.1.0"
coverage = "^5.3"
mypy = "^0.782"
teamcity-messages = "^1.28"
requests-mock = "^1.8.0"
pytest-describe = "^1.0.0"
faker = "^4.18.0"
freezegun = "^1.1.0"
twine = "^3.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
